requirement.txt
asttokens==2.4.0
colorama==0.4.6
executing==2.0.0
icecream==2.1.3
numpy==1.26.1
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
pydobot==1.3.2
Pygments==2.16.1
pyserial==3.4
pytesseract==0.3.10
six==1.16.0
